refactor(constants): Use namespace import internally
