+++ src/parser/classes/misc/Author.ts
@@ -1,4 +1,4 @@
-import Constants from '../../../utils/Constants.js';
+import * as Constants from '../../../utils/Constants.js';
 import type { YTNode} from '../../helpers.js';
 import { observe, type ObservedArray } from '../../helpers.js';
 import Parser, { type RawNode } from '../../index.js';
